% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_1.R
\name{figure_1}
\alias{figure_1}
\title{plots figure 1.}
\usage{
figure_1(x)
}
\arguments{
\item{x}{Input phase_date_data class object.}
}
\value{
Returns a ggplot object.
}
\description{
\code{year_phase_data} is the class used for the creation of all
figures and tables in the made up RAP report (it's a demo for the RAP MOOC),
including this bar chart.
}
\examples{

report_data <- DataTidying::phase_date_data(ƒ::regreg)
DataTidying::figure_1(report_data)

}
