% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_2.R
\name{figure_2}
\alias{figure_2}
\title{Figure 2 of the DataTidying RAP report.}
\usage{
figure_2(x)
}
\arguments{
\item{x}{Input object of \code{year_phase_data} class.}
}
\value{
Returns a ggplot object.
}
\description{
The function returns the time series ggplot2 object geom_line.
}
\examples{

library(DataTidying)
report_data <- phase_date_data(regreg)
figure_2(report_data)

}
