% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fivereg_recent.R
\name{fivereg_recent}
\alias{fivereg_recent}
\title{The five most recent published registers.}
\usage{
fivereg_recent(x, n = 5)
}
\arguments{
\item{x}{Input object of \code{year_phase_data} class.}

\item{n}{The \code{n} passed to \code{head}.}
}
\value{
Returns a character string of the five most recent registers.
}
\description{
The \code{fivereg_recent} function expects the
\code{year_phase_data} and outputs a character string.
}
\details{
The registers are sorted by publication date and name by
alphabetical order. The top five registers are output as a character string
including commas and an and, for inclusion in the report.
}
\examples{

library(regregrap)
report_data <- phase_date_data(regreg)
fivereg_recent(report_data)

}
