% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_date_data.R
\name{phase_date_data}
\alias{phase_date_data}
\title{minimal tidy data set for regregrap report production.}
\usage{
phase_date_data(
  x,
  log_level = futile.logger::WARN,
  log_appender = "console",
  eda = FALSE
)
}
\arguments{
\item{x}{Input dataframe, see details.}

\item{log_level}{The severity level at which log messages are written from
least to most serious: TRACE, DEBUG, INFO, WARN, ERROR, FATAL. Default is
level is INFO. See \code{?flog.threshold()} for additional details.}

\item{log_appender}{Defaults to write the log to "console", alternatively you
can provide a character string to specify a filename to also write to. See
for additional details \code{?futile.logger::appender.tee()}.}

\item{eda}{If TRUE an graphical data analysis is conducted for a human to check.}
}
\value{
If the class is not instantiated correctly, nothing is returned.
}
\description{
\code{year_phase_data} is the class used for the creation of all
figures and tables in the made up RAP report (it's a demo for the RAP MOOC).
}
\details{
The \code{year_phase_data} class expects a \code{data.frame} with at
least three columns: phase, date (entry-timestamp), and register name. Each
row represents a unique register.

Once inititated, the class has five slots: \code{df}: the basic
\code{data.frame}, \code{colnames}: a character vector containing the
column names from the \code{df}, \code{phase_levels}: a
factor vector containing levels of \code{df$phase} of the factor sector,
\code{yq}: an date vector containing \code{zoo::as.yearqrt(df$date)}.
}
\examples{

library(regregrap)

df <- phase_date_data(regreg)

}
